+++ changelog.md
@@ -1,4 +1,6 @@
 ### General
 
 - Added carpet rule `fluidNoPushPlayer`
+- Added carpet rule `slimeNoBounceUpPlayer`
+- Added carpet rule `kelpGrowOnlyIntoFullWater`
 - `betterOpPlayerNoCheat` now effects `/clone` as well